Platforms/RPi4: Add Trusted Firmware binaries
Similar to what is the case with the Raspberry Pi 3, the Raspberry Pi
4 UEFI firmware requires the provision of a Trusted Firmware binary
(TF-A).

The binary is built for a dtb base address of 0x00020000 and a UEFI
payload base address of 0x00030000.

Binaries are built using a custom version of TF-A (the source of which
is documented in the Readme), since the official repository is missing
a fix to get proper baudrate when using a recent version of start4.elf,
as we as the ability to switch between PL011 and mini UART. We are
currently working with TF-A to upstream both these fixes so that we can
switch to binaries built from official source at a later date.

In order to provide the ability to switch serial I/O controllers, a
feature we will be adding to the RPi4 UEFI Firmware, and not lose TF-A
output, two binaries are provided, one for mini UART output and one for
PL011 output.

Signed-off-by: Pete Batard <pete@akeo.ie>
Reviewed-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>