Platform/RPi/TrustedFirmware: Update TF-A binaries to v2.3 release
With the v2.3 release of TF-A, that integrates automatic PL011 vs miniUART
switching, we can remove the dual bl31 binaries we were using for RPi4,
and of course, we also take this opportunity to update the RPi3 binaries
to the latest.

Note that these binaries were built in an open and verifiable manner
through an AppVeyor build script (https://github.com/pbatard/pitf).

Signed-off-by: Pete Batard <pete@akeo.ie>
Reviewed-by: Ard Biesheuvel <ard.biesheuvel@arm.com>