KabylakeOpenBoardBinPkg/AspireVn7Dash572G: Add VBT
Add VBT for the Acer Aspire VN7-572G board. Changes were ported from
the vendor's VBT (version 200) to VBT version 221 (version 228 changed
the units for some eDP settings). Most of the changes were to display
initialisation timings, which were corroborated with the datasheet.

A display toggle table may be implemented in the future.

Cc: Chasel Chiu <chasel.chiu@intel.com>
Cc: Nate DeSimone <nathaniel.l.desimone@intel.com>
Cc: Isaac Oram <isaac.w.oram@intel.com>
Cc: Michael Kubacki <michael.kubacki@microsoft.com>
Signed-off-by: Benjamin Doron <benjamin.doron00@gmail.com>
Reviewed-by: Nate DeSimone <nathaniel.l.desimone@intel.com>