Platform/O6N: iomask two USB3.0 ports VBUS
When booting Radxa OS[0] from any one of the USB3.0 ports using a 10Gbps
SSD or SATA disk USB case, it's possible that the systemd bootup process
hangs, and then, after a few minutes, the serial console outputs a large
number of I/O error messages. Just like the following:
```
[pid:308,cpu4,scsi_eh_1]I/O error, dev sdb, sector 3687456 op 0x0:(READ) flags 0x84700 phys_seg 128 prio class 3
[pid:241,cpu4,kworker/4:1H]sd 1:0:0:0: rejecting I/O to offline device
[pid:241,cpu4,kworker/4:1H]I/O error, dev sdb, sector 9068544 op 0x0:(READ) flags 0x83700 phys_seg 1 prio class 2
[pid:308,cpu1,scsi_eh_1]I/O error, dev sdb, sector 3688480 op 0x0:(READ) flags 0x84700 phys_seg 128 prio class 3
[pid:241,cpu4,kworker/4:1H]I/O error, dev sdb, sector 9068648 op 0x0:(READ) flags 0x83700 phys_seg 1 prio class 2
...
```

It's found that the USB3.0 ports' VBUS pins: USB_DRIVE_VBUS4 and
USB_DRIVE_VBUS5, have a voltage drop and increase during the kernel
initialisation; this is supposed to be the reason for the issue mentioned
above.

To avoid the voltage variations on these two pins during the bootup. We
set the functions of them to GPIO in Iomux.asl and PlatformEnvHookLib.c,
and include them in the `PcdAcpiGpio4IoMask`.

Because these two pins are configured as GPIO. They should be configured
in ArmPsciResetSystemLib.c so the USB3.0 ports will be disabled after
powering off.

[0]: https://github.com/radxa-build/orion-o6/releases/tag/rsdk-t2
[1]: https://github.com/radxa/edk2-platforms/blob/efacf52b4a2bf441c7aa37054f76623a76961bf9/Platform/Radxa/Orion/O6N/O6N.dsc#L406-L407

Signed-off-by: Zhaoming Luo <luozhaoming@radxa.com>

Origin-Repo: radxa/edk2-platforms
Origin-Commit: 5e7faa35e2e894da4a9fad7e164a439fe22e60ee
Canonical-Upstream: tianocore/edk2-platforms
Upstream-Status: vendor-delta
Impact-Area: meta
Output-Impact: unknown